+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - added arithmetization oriented authenticated encryption with associated data (AEAD) scheme ([#480](https://github.com/0xMiden/crypto/pull/480)).
 - Added XChaCha20-Poly1305 AEAD scheme ([#484](https://github.com/0xMiden/crypto/pull/484)).
-- [BREAKING] `SmtLeaf::entries()` now returns an iterator ([#521](https://github.com/0xMiden/crypto/pull/521)).
+- [BREAKING] `SmtLeaf::entries()` now returns a slice ([#521](https://github.com/0xMiden/crypto/pull/521)).
 
 ## 0.16.1 (2025-08-21)